requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
marshmallow==3.10.0
mypy==0.800
mypy-extensions==0.4.3
pathspec==0.8.1
python-dateutil==2.8.1
regex==2020.11.13
simplejson==3.17.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
